CA_GetRoomCenter
Cast a ray in three different directions, do some weird math, get center of the "room". This returns the radius of the circle formed from the center, and -1 when something goes wrong.

This will work for square and SOME rectangles. It will also work for circular rooms.

Rectangles: If the rays only hit two walls, this won't work. If the rays hit 3 walls it will.
Circles: It will ALWAYS work. Unless for some reason one of the rays don't hit an object.